Implemented the kinetic energy density in momentum space
The K.E.D. in momentum space is available through the program
dtkmomd by activating the option -p K.
Also, in this commit the file DESIGNTODOS is added to the
track list. The file DESIGNTODOS contains some ideas for
recoding dtk, with the purpose of making it even more
object-oriented.