critPtNetWork redesigned. 4h20-anion correctly processed.
the class critPtNetWork was redesigned. Some functions
were simply copied from the old version.

This class now correctly find all cps for the water anion complex.

Some files required small changes in order to make them
compatible with the new design. The major change for this was
about the inclusion of a gaussWaveFunction pointer to
the constructor of critPtNetWork. With this pointer,
there is no need to frequently pass by reference the
wave function object.