Implemented RoSE, LED, and RedDensGrad (s)
The implementation of the fields RoSe, LED, and s (RedDensGrad)
in wavefunctionclass.* is done in this commit.

The functions displayAllFieldProperties and writeAllFieldProperties
also includes the information of these new fields. The magnitudes of
GradRho, and GradLOL are also included, as well as the
eigenvectors of HessLOL.

The file fldtypesdef.h is modified to include these fields.