First implementation of RingPath search.
The function does not work for every case, only for the
simplest cases. Basic output is implemented in order to save
the information as well as plot the results using povray.
The search of ring paths is only implemented in
testrwf, but not on dtkfindcp (the function is not completely
ready).