Make integrator_vegas to be a derived class of integrator3d.
After this commit, IntegratorVegas is a derived class of
Integrator3D, which implies that the integrand is provided
through a Function3D pointer, as opposed to be included
in IntegratorVegas. This, in principle should allow, in
the future, to re-use IntegratorVegas without DTK context.

In addition, IntegratorVegas and Integrator3DMiser and
other integrators to be added to dtkintegrate.