Fixbug: Orca wfx and wfn files were not read correctly.
Although the wfx format requires tags to be printed in
unique single lines, this fix allows tags to be read from
lines including additional information.

In the previous version, the MO number was not implemented
properly. The format used in proaim is:
format (35x,f12.8,15x,f12.8)
i.e., the orbital energy starts at char 62 (not 63!).
In addition, the fix now forces to advance the buffer pos
(after readin the MO's energy) 75 records (74chars+1LF).
This might cause problems for windows-style Line endings;
yet we will not solve this issue in this fix.