Fixbugs: EvalDensityMatrix1, EvalDensity, and EvalFTRho2ClosedShell (parallel)
Parallel versions had some numerical errors in the parallel versions.
The errors in EvalDensityMatrix1 and EvalDensity were introduced in the
parallel further-optimization commits. It were not caught, until the
XeF2.wfx was used. XeF2.wfx has EDFs, but no single-spin orbitals.
The bug in EvalFTRho2ClosedShell was not exactly a bug, since the
previous commit of this function did not have the correct parallel
version.